requirements.txt
blinker==1.9.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
Flask==3.1.3
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
requests==2.33.1
urllib3==2.6.3
Werkzeug==3.1.8
